+++ front/components/assistant/AssistantDropdownMenu.tsx
